# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Cloud Storage API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface AuthStatus {
  authenticated: boolean;
  /** @nullable */
  username: string | null;
}

export interface LoginInput {
  password: string;
}

export type FileItemType = typeof FileItemType[keyof typeof FileItemType];


export const FileItemType = {
  file: 'file',
  folder: 'folder',
} as const;

export interface FileItem {
  name: string;
  path: string;
  type: FileItemType;
  size: number;
  modifiedAt: string;
  /** @nullable */
  mimeType: string | null;
}

export interface Breadcrumb {
  name: string;
  path: string;
}

export interface FileList {
  path: string;
  items: FileItem[];
  breadcrumbs: Breadcrumb[];
}

export interface MoveInput {
  from: string;
  to: string;
}

export interface FolderInput {
  path: string;
}

export interface UploadResult {
  uploaded: number;
  files: FileItem[];
}

export interface ShareInput {
  path: string;
  /** @nullable */
  expiresAt: string | null;
}

export interface ShareLink {
  token: string;
  url: string;
  /** @nullable */
  expiresAt: string | null;
}

export interface ShareInfo {
  token: string;
  filename: string;
  /** @nullable */
  mimeType: string | null;
  size: number;
  /** @nullable */
  expiresAt: string | null;
}

export interface StorageStats {
  totalFiles: number;
  totalFolders: number;
  totalSize: number;
  largestFiles: FileItem[];
}

export type ListFilesParams = {
path?: string;
search?: string;
sort?: ListFilesSort;
order?: ListFilesOrder;
};

export type ListFilesSort = typeof ListFilesSort[keyof typeof ListFilesSort];


export const ListFilesSort = {
  name: 'name',
  size: 'size',
  date: 'date',
} as const;

export type ListFilesOrder = typeof ListFilesOrder[keyof typeof ListFilesOrder];


export const ListFilesOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;

export type DeleteFileParams = {
path: string;
};

export type UploadFilesParams = {
path?: string;
};