# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Personal Cloud Storage API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: files
    description: File management
  - name: share
    description: Share links
  - name: storage
    description: Storage statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current auth status
      responses:
        "200":
          description: Auth status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"

  /files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List files in a directory
      parameters:
        - name: path
          in: query
          schema:
            type: string
            default: "/"
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, size, date]
            default: name
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        "200":
          description: File listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /file:
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file or folder
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: moveFile
      tags: [files]
      summary: Rename or move a file or folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveInput"
      responses:
        "200":
          description: Moved/renamed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /folders:
    post:
      operationId: createFolder
      tags: [files]
      summary: Create a new folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FolderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /upload:
    post:
      operationId: uploadFiles
      tags: [files]
      summary: Upload files (multipart/form-data, handled client-side with fetch)
      parameters:
        - name: path
          in: query
          schema:
            type: string
            default: "/"
      responses:
        "200":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /share:
    post:
      operationId: createShare
      tags: [share]
      summary: Create a shareable link for a file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShareInput"
      responses:
        "201":
          description: Share link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareLink"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /share/{token}:
    get:
      operationId: getShare
      tags: [share]
      summary: Get share link info (public)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Share info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareInfo"
        "404":
          description: Not found or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /storage/stats:
    get:
      operationId: getStorageStats
      tags: [storage]
      summary: Get storage usage statistics
      responses:
        "200":
          description: Storage stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    AuthStatus:
      type: object
      properties:
        authenticated:
          type: boolean
        username:
          type: ["string", "null"]
      required:
        - authenticated
        - username

    LoginInput:
      type: object
      properties:
        password:
          type: string
      required:
        - password

    FileItem:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, folder]
        size:
          type: integer
        modifiedAt:
          type: string
          format: date-time
        mimeType:
          type: ["string", "null"]
      required:
        - name
        - path
        - type
        - size
        - modifiedAt
        - mimeType

    FileList:
      type: object
      properties:
        path:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/FileItem"
        breadcrumbs:
          type: array
          items:
            $ref: "#/components/schemas/Breadcrumb"
      required:
        - path
        - items
        - breadcrumbs

    Breadcrumb:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
      required:
        - name
        - path

    MoveInput:
      type: object
      properties:
        from:
          type: string
        to:
          type: string
      required:
        - from
        - to

    FolderInput:
      type: object
      properties:
        path:
          type: string
      required:
        - path

    UploadResult:
      type: object
      properties:
        uploaded:
          type: integer
        files:
          type: array
          items:
            $ref: "#/components/schemas/FileItem"
      required:
        - uploaded
        - files

    ShareInput:
      type: object
      properties:
        path:
          type: string
        expiresAt:
          type: ["string", "null"]
          format: date-time
      required:
        - path
        - expiresAt

    ShareLink:
      type: object
      properties:
        token:
          type: string
        url:
          type: string
        expiresAt:
          type: ["string", "null"]
          format: date-time
      required:
        - token
        - url
        - expiresAt

    ShareInfo:
      type: object
      properties:
        token:
          type: string
        filename:
          type: string
        mimeType:
          type: ["string", "null"]
        size:
          type: integer
        expiresAt:
          type: ["string", "null"]
          format: date-time
      required:
        - token
        - filename
        - mimeType
        - size
        - expiresAt

    StorageStats:
      type: object
      properties:
        totalFiles:
          type: integer
        totalFolders:
          type: integer
        totalSize:
          type: integer
        largestFiles:
          type: array
          items:
            $ref: "#/components/schemas/FileItem"
      required:
        - totalFiles
        - totalFolders
        - totalSize
        - largestFiles