# lib/api-zod/src/generated/types/fileItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Cloud Storage API
 * OpenAPI spec version: 0.1.0
 */
import type { FileItemType } from './fileItemType';

export interface FileItem {
  name: string;
  path: string;
  type: FileItemType;
  size: number;
  modifiedAt: Date;
  /** @nullable */
  mimeType: string | null;
}