# lib/api-zod/src/generated/types/fileList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Cloud Storage API
 * OpenAPI spec version: 0.1.0
 */
import type { Breadcrumb } from './breadcrumb';
import type { FileItem } from './fileItem';

export interface FileList {
  path: string;
  items: FileItem[];
  breadcrumbs: Breadcrumb[];
}