# lib/api-zod/src/generated/types/shareInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Cloud Storage API
 * OpenAPI spec version: 0.1.0
 */

export interface ShareInfo {
  token: string;
  filename: string;
  /** @nullable */
  mimeType: string | null;
  size: number;
  /** @nullable */
  expiresAt: Date | null;
}