# lib/api-zod/src/generated/types/shareInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Cloud Storage API
 * OpenAPI spec version: 0.1.0
 */

export interface ShareInput {
  path: string;
  /** @nullable */
  expiresAt: Date | null;
}