# lib/api-zod/src/generated/types/shareLink.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Cloud Storage API
 * OpenAPI spec version: 0.1.0
 */

export interface ShareLink {
  token: string;
  url: string;
  /** @nullable */
  expiresAt: Date | null;
}