# lib/api-zod/src/generated/types/storageStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Cloud Storage API
 * OpenAPI spec version: 0.1.0
 */
import type { FileItem } from './fileItem';

export interface StorageStats {
  totalFiles: number;
  totalFolders: number;
  totalSize: number;
  largestFiles: FileItem[];
}